refactor(RateLimiter): pack Bucket struct from 4 slots to 2
Narrow capacity, refillPerSecond, and remainingAmount from uint256 to
uint128, co-locating config fields in one slot and state fields in
another. This halves SLOADs in _enforceRateLimit and reduces SSTOREs
from 2 to 1 on the hot path. Use UIntMath.safe128 for the uint256→uint128
downcast in _enforceRateLimit so oversized amounts revert with a clear
InvalidUInt128 error before reaching the bucket check.

+++ src/abstract/interfaces/IRateLimiter.sol
@@ -9,12 +9,12 @@ interface IRateLimiter {
     /// @notice Reverted when a mint would exceed the available rate limit.
     /// @param  requested Amount requested.
     /// @param  available Amount available in bucket.
-    error RateLimitExceeded(uint256 requested, uint256 available);
+    error RateLimitExceeded(uint128 requested, uint128 available);
 
     /// @notice Reverted when removing config with non-zero params.
     /// @param  capacity        Invalid non-zero capacity.
     /// @param  refillPerSecond Invalid non-zero refill rate.
-    error InvalidRateLimitRemoval(uint256 capacity, uint256 refillPerSecond);
+    error InvalidRateLimitRemoval(uint128 capacity, uint128 refillPerSecond);
 
     /// @notice Reverted when rate limit manager is the zero address.
     error ZeroRateLimitManager();
@@ -26,17 +26,17 @@ interface IRateLimiter {
     /// @param  capacity        Maximum bucket capacity.
     /// @param  refillPerSecond Refill rate per second.
     /// @param  enabled         True when configured/updated, false when removed.
-    event RateLimitSet(address indexed issuer, uint256 capacity, uint256 refillPerSecond, bool enabled);
+    event RateLimitSet(address indexed issuer, uint128 capacity, uint128 refillPerSecond, bool enabled);
 
     /* ============ Interactive Functions ============ */
 
     /// @notice Sets/removes issuer mint rate-limit configuration.
-    /// @dev    MUST only be callable by RATE_LIMITER_ROLE.
+    /// @dev    MUST only be callable by RATE_LIMIT_MANAGER_ROLE.
     /// @param  issuer          Address of the issuer.
     /// @param  capacity        Maximum bucket capacity.
     /// @param  refillPerSecond Refill rate per second. Set to 0 for a one-time mint cap with no refill.
     /// @param  enabled         True to configure/update. False to remove (requires zero capacity/refillPerSecond).
-    function setRateLimit(address issuer, uint256 capacity, uint256 refillPerSecond, bool enabled) external;
+    function setRateLimit(address issuer, uint128 capacity, uint128 refillPerSecond, bool enabled) external;
 
     /* ============ View/Pure Functions ============ */
 
@@ -47,10 +47,10 @@ interface IRateLimiter {
     /// @param  issuer          Address of the issuer.
     /// @return capacity        Maximum bucket capacity.
     /// @return refillPerSecond Refill rate per second.
-    function getRateLimitConfig(address issuer) external view returns (uint256 capacity, uint256 refillPerSecond);
+    function getRateLimitConfig(address issuer) external view returns (uint128 capacity, uint128 refillPerSecond);
 
     /// @notice Returns currently available amount in issuer mint bucket (after virtual refill).
     /// @param  issuer Address of the issuer.
     /// @return The remaining amount in the bucket.
-    function getRemainingAmount(address issuer) external view returns (uint256);
+    function getRemainingAmount(address issuer) external view returns (uint128);
 }
